úprava auth backendu
