+++ backend/routes/auth/services/password.js
@@ -1,0 +1,16 @@
+/* eslint-disable */
+const bcrypt = require('bcrypt');
+
+const hashPassword = async (password) => {
+  const salt = await bcrypt.genSalt(10);
+  return await bcrypt.hash(password, salt);
+};
+
+const comparePassword = async (plainPassword, hashedPassword) => {
+  return await bcrypt.compare(plainPassword.trim(), hashedPassword);
+};
+
+module.exports = {
+  hashPassword,
+  comparePassword,
+};
