+++ backend/routes/auth/signup.js
@@ -1,0 +1,67 @@
+/* eslint-disable */
+const express = require('express');
+const router = express.Router();
+const pool = require('../../config/database');
+const { hashPassword } = require('./services/password');
+const { validateSignupData } = require('./services/validation');
+
+router.post('/signup', async (req, res) => {
+  const client = await pool.connect();
+  try {
+    const { email, password, firstName, lastName } = req.body;
+
+    // Validate input
+    const validation = validateSignupData(email, password, firstName, lastName);
+    if (!validation.isValid) {
+      return res.status(400).json({ error: validation.errors.join(', ') });
+    }
+
+    // Start transaction
+    await client.query('BEGIN');
+
+    // Check if user already exists
+    const existingUser = await client.query('SELECT id FROM users WHERE email = $1', [email]);
+    if (existingUser.rows.length > 0) {
+      await client.query('ROLLBACK');
+      return res.status(400).json({ error: 'Email already registered' });
+    }
+
+    // Hash password
+    const hashedPassword = await hashPassword(password);
+
+    // Insert new user
+    const insertQuery = `
+      INSERT INTO users (email, password, firstname, surname, role)
+      VALUES ($1, $2, $3, $4, $5)
+      RETURNING id, email, firstname, surname, role`;
+
+    const result = await client.query(insertQuery, [
+      email.toLowerCase(),
+      hashedPassword,
+      firstName,
+      lastName,
+      'parent',
+    ]);
+
+    await client.query('COMMIT');
+
+    res.status(201).json({
+      message: 'User registered successfully',
+      user: {
+        id: result.rows[0].id,
+        email: result.rows[0].email,
+        firstname: result.rows[0].firstname,
+        surname: result.rows[0].surname,
+        role: result.rows[0].role,
+      },
+    });
+  } catch (err) {
+    await client.query('ROLLBACK');
+    console.error('Signup error:', err);
+    res.status(500).json({ error: 'Registration failed', details: err.message });
+  } finally {
+    client.release();
+  }
+});
+
+module.exports = router;
